.babelrc파일 추가
