+++ public/index.html
@@ -1,0 +1,107 @@
+<!DOCTYPE html>
+<html lang="ko">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>판다마켓</title>
+    <meta property="og:title" content="판다 마켓" />
+    <meta property="og:description" content="링크 설명 입력" />
+    <meta property="og:image" content="Image/미리보기_이미지.png" />
+    <link rel="stylesheet" href="../styles/style.css" />
+  </head>
+  <body>
+    <div class="header-frame">
+      <a href="/"><img id="logo" src="Image/logo.png" alt="판다마켓" /></a>
+      <a href="login.html"
+        ><img id="login" src="Image/Img_login.png" alt="로그인 버튼"
+      /></a>
+    </div>
+
+    <div class="header">
+      <div class="header-description">
+        <h1>
+          일상의 모든 물건을<br />
+          거래해 보세요
+        </h1>
+        <a href="/items"
+          ><img
+            id="header_click"
+            src="Image/Img_header_click.png"
+            alt="구경하러 가기"
+        /></a>
+      </div>
+      <img id="home_top" src="Image/Img_home_top.png" />
+    </div>
+    <div class="section" id="wrapper">
+      <div class="section-box">
+        <img id="home_01" src="Image/Img_home_01.png" />
+        <div class="section-description">
+          <h2 class="section-tag">HOT ITEM</h2>
+          <h1>
+            인기 상품을 <br />
+            확인해 보세요
+          </h1>
+          <p class="section-small">
+            가장 HOT한 중고거래 물품을<br />
+            판다마켓에서 확인해 보세요.
+          </p>
+        </div>
+      </div>
+      <div class="section-box" id="section2">
+        <div class="section-description">
+          <div class="section-description">
+            <h2 class="section-tag">SEARCH</h2>
+            <h1>
+              구매를 원하는<br />
+              상품을 검색하세요
+            </h1>
+            <p class="section-small">
+              구매하고 싶은 물품은 검색해서<br />
+              쉽게 찾아보세요.
+            </p>
+          </div>
+        </div>
+        <img id="home_02" src="Image/Img_home_02.png" />
+      </div>
+      <div class="section-box" id="section3">
+        <img id="home_03" src="Image/Img_home_03.png" />
+        <div class="section-description">
+          <h2 class="section-tag">Register</h2>
+          <h1>
+            판매를 원하는<br />
+            상품을 등록하세요
+          </h1>
+          <p class="section-small">
+            어떤 물건이든 판매하고 싶은 상품을 쉽게 등록하세요.
+          </p>
+        </div>
+      </div>
+    </div>
+
+    <div class="section4">
+      <div class="section4-description">
+        믿을 수 있는<br />
+        판다마켓 중고거래
+      </div>
+      <img id="home_bottom" src="Image/Img_home_bottom.png" />
+    </div>
+    <footer>
+      <div class="codeit">©codeit - 2024</div>
+      <div class="faq">Privacy Policy FAQ</div>
+      <div class="icons">
+        <a href="https://www.facebook.com/?locale=ko_KR" target="_blank"
+          ><img id="facebook" src="Image/ic_facebook.png" alt="페이스북"
+        /></a>
+        <a href="https://twitter.com/?lang=ko" target="_blank"
+          ><img id="x" src="Image/ic_twitter.png" alt="트위터"
+        /></a>
+        <a href="https://www.youtube.com/" target="_blank"
+          ><img id="youtube" src="Image/ic_youtube.png" alt="유튜브"
+        /></a>
+        <a href="https://www.instagram.com/" target="_blank"
+          ><img id="insta" src="Image/ic_instagram.png" alt="인스타"
+        /></a>
+      </div>
+    </footer>
+  </body>
+</html>
